refactor(bottom-border): add BottomBorderContainer widget and integrate into Dashboard and Settings screens
