+++ lib/desktop/features/dashboard/presentation/screens/dashboard_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:streamkeys/core/constants/colors.dart';
+import 'package:streamkeys/common/widgets/bottom_border_container.dart';
 import 'package:streamkeys/common/widgets/screen_tab_bar.dart';
 import 'package:streamkeys/common/widgets/page_tab.dart';
 
@@ -36,8 +36,7 @@ class _DashboardScreenState extends State<DashboardScreen>
     return Scaffold(
       appBar: PreferredSize(
         preferredSize: const Size.fromHeight(45),
-        child: Container(
-          decoration: _boxDecoration(context),
+        child: BottomBorderContainer(
           child: ScreenTabBar(
             tabs: widget.tabs,
             controller: _tabController,
@@ -51,14 +50,5 @@ class _DashboardScreenState extends State<DashboardScreen>
     );
   }
 
-  BoxDecoration _boxDecoration(BuildContext context) {
-    return BoxDecoration(
-      border: Border(
-        bottom: BorderSide(
-          width: 4,
-          color: AppColors.of(context).outlineVariant,
-        ),
-      ),
-    );
-  }
+  
 }
